review-analyser/pyproject.toml
[project]
name = "review-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
